requirements.txt
torch==1.4.0
dgl-cu101==0.4.3.post2
scipy==1.1.0
numpy==1.15.4
tqdm==4.42.1
networkx>=2.1
